+++ src/pymedia/main.py
@@ -18,11 +18,11 @@ if sys.platform == "win32" and isinstance(sys.stdout, io.TextIOWrapper):
 
 from pymedia.typer.commands.animated_command import animated_typer
 from pymedia.typer.commands.audio_commands import audio_typer
+from pymedia.typer.commands.cut_command import cut_typer
 from pymedia.typer.commands.info_command import info_typer
 from pymedia.typer.commands.join_command import join_typer
 from pymedia.typer.commands.remux_command import remux_typer
 from pymedia.typer.commands.sheet_command import sheet_typer
-from pymedia.typer.commands.split_command import split_typer
 from pymedia.typer.commands.subtitles_commands import subtitles_typer
 from pymedia.typer.commands.thumb_commands import thumb_typer
 from pymedia.typer.commands.transcode_command import transcode_typer
@@ -31,10 +31,10 @@ from pymedia.typer.instance import typer_instance
 app = typer_instance
 app.add_typer(info_typer)
 app.add_typer(sheet_typer)
-app.add_typer(join_typer)
-app.add_typer(remux_typer)
-app.add_typer(split_typer)
 app.add_typer(transcode_typer)
+app.add_typer(remux_typer)
+app.add_typer(cut_typer)
+app.add_typer(join_typer)
 app.add_typer(audio_typer)
 app.add_typer(subtitles_typer)
 app.add_typer(animated_typer)