+++ src/pymedia/typer/commands/cut_command.py
@@ -1,10 +1,11 @@
-"""Comando Typer para iniciar la división de contenedores."""
+"""Comando Typer para iniciar el corte de contenedores."""
 
 import typer
 
-from pymedia.errors import MissingArgumentError
-from pymedia.pipeline.split_pipeline import SplitPipeline
-from pymedia.typer.help import SPLIT_HELP
+from pymedia.errors import MissingRequiredOptionsError, UserError
+from pymedia.locales import _  # noqa
+from pymedia.pipeline.cut_pipeline import CutPipeline
+from pymedia.typer.help import CUT_HELP
 from pymedia.typer.options import (
     DebugOption,
     HelpOption,
@@ -12,21 +13,25 @@ from pymedia.typer.options import (
     OutputOption,
     OverwriteOption,
     TimestampAtMediaOption,
+    TimestampEndMediaOption,
+    TimestampStartMediaOption,
 )
 from pymedia.types import OverwriteMode
 
-split_typer = typer.Typer()
+cut_typer = typer.Typer()
 
 
-@split_typer.command(
-    name="split",
-    help=SPLIT_HELP,
+@cut_typer.command(
+    name="cut",
+    help=CUT_HELP,
     rich_help_panel="Video commands",
     no_args_is_help=True,
 )
-def split(
+def cut(
     media_input: MediaInputArgument,
-    timestamp_at: TimestampAtMediaOption,
+    timestamp_at: TimestampAtMediaOption = None,
+    timestamp_start: TimestampStartMediaOption = None,
+    timestamp_end: TimestampEndMediaOption = None,
     media_output: OutputOption = None,
     overwrite: OverwriteOption = OverwriteMode.ASK,
     debug: DebugOption = False,
@@ -37,6 +42,8 @@ def split(
     Args:
         media_input: Ruta del fichero de vídeo a procesar.
         timestamp_at: Lista de marcas de tiempo para dividir el vídeo.
+        timestamp_start: Indica cuando empieza el vídeo de salida.
+        timestamp_end: Indica cuando termina el vídeo de salida.
         media_output: Ruta absoluta del fichero de salida procesado.
         overwrite: Política ante conflicto de salida ya existente.
         debug: Habilita el nivel de log DEBUG.
@@ -45,13 +52,22 @@ def split(
     Raises:
         MissingArgumentError: Si no se indica la lista de marcas de tiempo.
     """
-    if timestamp_at is None:
-        raise MissingArgumentError(name="timestamp_at")
+    if timestamp_at is None and timestamp_start is None and timestamp_end is None:
+        raise MissingRequiredOptionsError(options=["at", "start", "end"])
 
-    pipeline = SplitPipeline(debug=debug)
+    if timestamp_at is not None and (
+        timestamp_start is not None or timestamp_end is not None
+    ):
+        raise UserError(
+            msg=_("Ambiguous options: You can't select --at with --start or --end.")
+        )
+
+    pipeline = CutPipeline(debug=debug)
     pipeline.process_parameters(
         media_input=media_input,
         timestamp_at=timestamp_at,
+        timestamp_start=timestamp_start,
+        timestamp_end=timestamp_end,
         media_output=media_output,
         overwrite=overwrite,
     )